+++ src/viewpoint-measures/I3.cpp
@@ -17,11 +17,11 @@ I3::I3(const QString &pName): Measure(pName, true)
 
 void I3::Compute(const SceneInformationBuilder *pSceneInformationBuilder)
 {
-    std::vector< int > elementsOutOfDomain;
+    std::vector< size_t > elementsOutOfDomain;
 
     const auto projectedAreasMatrix = pSceneInformationBuilder->GetProjectedAreasMatrix();
-    int numberOfViewpoints = projectedAreasMatrix->GetNumberOfViewpoints();
-    int numberOfPolygons = projectedAreasMatrix->GetNumberOfPolygons();
+    size_t numberOfViewpoints = projectedAreasMatrix->GetNumberOfViewpoints();
+    size_t numberOfPolygons = projectedAreasMatrix->GetNumberOfPolygons();
 
     mValues.resize( numberOfViewpoints );
     std::fill(mValues.begin(), mValues.end(), 0.0f);
@@ -29,7 +29,7 @@ void I3::Compute(const SceneInformationBuilder *pSceneInformationBuilder)
     //The polygonal I2 is computed first
     std::vector< float > polygonalI2(numberOfPolygons, 0.0f);
     unsigned int sum_a_t = projectedAreasMatrix->GetTotalSum();
-    for( int currentPolygon = 0; currentPolygon < numberOfPolygons; currentPolygon++ )
+    for( size_t currentPolygon = 0; currentPolygon < numberOfPolygons; currentPolygon++ )
     {
         unsigned int sum_a_z = projectedAreasMatrix->GetSumPerPolygon(currentPolygon);
         float sumAux1 = 0.0f;
@@ -56,7 +56,7 @@ void I3::Compute(const SceneInformationBuilder *pSceneInformationBuilder)
         unsigned int a_t = projectedAreasMatrix->GetSumPerViewpoint(currentViewpoint);
         if( a_t != 0 )
         {
-            for( int currentPolygon = 0; currentPolygon < numberOfPolygons; currentPolygon++ )
+            for( size_t currentPolygon = 0; currentPolygon < numberOfPolygons; currentPolygon++ )
             {
                 unsigned int a_z = projectedAreasMatrix->GetValue(currentViewpoint, currentPolygon);
                 if( a_z != 0)
@@ -76,7 +76,7 @@ void I3::Compute(const SceneInformationBuilder *pSceneInformationBuilder)
         }
     }
     //The minimum value is assigned to the viewpoints out of the domain (viewpoints that don't see anything)
-    for( int currentElement = 0; currentElement < elementsOutOfDomain.size(); currentElement++ )
+    for( size_t currentElement = 0; currentElement < elementsOutOfDomain.size(); currentElement++ )
     {
         mValues[elementsOutOfDomain.at(currentElement)] = minValue;
     }
