+++ src/viewpoint-measures/ProjectedArea.cpp
@@ -12,7 +12,7 @@ ProjectedArea::ProjectedArea(const QString &pName): Measure(pName, true)
 void ProjectedArea::Compute(const SceneInformationBuilder *pSceneInformationBuilder)
 {
     const auto projectedAreasMatrix = pSceneInformationBuilder->GetProjectedAreasMatrix();
-    int numberOfViewpoints = projectedAreasMatrix->GetNumberOfViewpoints();
+    size_t numberOfViewpoints = projectedAreasMatrix->GetNumberOfViewpoints();
     mValues.resize( numberOfViewpoints );
     std::fill(mValues.begin(), mValues.end(), 0.0f);
     for( size_t currentViewpoint = 0; currentViewpoint < numberOfViewpoints; currentViewpoint++ )
